more logging when creating test users

+++ cristin-commons/src/test/java/no/unit/nva/utils/CognitoUtilCreateUsers.java
@@ -3,6 +3,10 @@ package no.unit.nva.utils;
 import no.unit.nva.cognito.CognitoUtil;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import static java.util.Objects.isNull;
 
 @Tag("integrationTest")
 @Tag("createTestUsers")
@@ -15,14 +19,20 @@ public class CognitoUtilCreateUsers {
     private static final String USER_POOL_ID = "eu-west-1_DNRmDPtxY";
     private static final String CLIENT_APP_ID = "4qfhv3kl9qcr2knsfb8lhu1u40";
 
+    private static final Logger logger = LoggerFactory.getLogger(CognitoUtilCreateUsers.class);
+
     private static final CognitoUtil cognitoUtil = new CognitoUtil(USER_POOL_ID, CLIENT_APP_ID, REGION);
 
     @Test
 
     void createTestUser() {
-        System.out.println("createTestUser");
-        cognitoUtil.deleteUser(FEIDE_ID);
-        cognitoUtil.createUser(FEIDE_ID, PASSWORD);
+        if (isNull(cognitoUtil.loginUser(FEIDE_ID, PASSWORD))) {
+            logger.info("createTestUser");
+            cognitoUtil.deleteUser(FEIDE_ID);
+            cognitoUtil.createUser(FEIDE_ID, PASSWORD);
+        } else {
+            logger.info("testuser not changed, exists and can login");
+        }
     }
 
 }
